Changed --debugger-view into --report=debugger, and moved it into the core library. Made the output look more like the other outputs of the tool.
