+++ llvm/tools/llvm-debuginfo-analyzer/llvm-debuginfo-analyzer.cpp
@@ -11,7 +11,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "DebuggerView.h"
 #include "Options.h"
 #include "llvm/DebugInfo/LogicalView/Core/LVOptions.h"
 #include "llvm/DebugInfo/LogicalView/LVReaderHandler.h"
@@ -96,10 +95,9 @@ int main(int argc, char **argv) {
   cl::extrahelp HelpResponse(
       "\nPass @FILE as argument to read options from FILE.\n");
 
-  cl::HideUnrelatedOptions({&AttributeCategory, &CompareCategory,
-                            &InternalCategory, &OutputCategory, &PrintCategory,
-                            &ReportCategory, &SelectCategory, &WarningCategory,
-                            &debuggerview::Category});
+  cl::HideUnrelatedOptions(
+      {&AttributeCategory, &CompareCategory, &InternalCategory, &OutputCategory,
+       &PrintCategory, &ReportCategory, &SelectCategory, &WarningCategory});
   cl::ParseCommandLineOptions(argc, argv,
                               "Printing a logical representation of low-level "
                               "debug information.\n");
@@ -122,11 +120,8 @@ int main(int argc, char **argv) {
     llvm::append_range(Objects, Objs);
   }
 
-  if (debuggerview::Enable)
-    return debuggerview::printDebuggerView(Objects, OutputFile.os());
-
-  ScopedPrinter W(OutputFile.os());
   propagateOptions();
+  ScopedPrinter W(OutputFile.os());
   LVReaderHandler ReaderHandler(Objects, W, ReaderOptions);
 
   // Print the command line.
